+++ packages/eyes-cypress/src/plugin/runningTests.js
@@ -8,6 +8,16 @@ const runningTests = {
   reset() {
     this.tests = [];
   },
+  async abortTests() {
+    return Promise.all(this.tests.map(async (test) => {
+      if (test.closePromise) {
+        // this condition doesn't really happen in the code, but it's here for purity. This is used in handlers' batchStart in order to cleanup the state in case a zombie open was called.
+        await test.closePromise
+      } else {
+        await test.abort()
+      }
+    }))
+  }
 };
 
 module.exports = runningTests;